+++ lib/src/http/browser/utils.dart
@@ -18,14 +18,13 @@ import 'dart:html';
 import 'package:w_transport/src/http/request_progress.dart';
 
 /// Transforms an [ProgressEvent] stream from an [HttpRequest] into
-/// a [WProgress] stream.
+/// a [RequestProgress] stream.
 StreamTransformer<ProgressEvent, RequestProgress> transformProgressEvents =
     StreamTransformer<ProgressEvent, RequestProgress>(
         (Stream<ProgressEvent> input, bool cancelOnError) {
-  late StreamController<RequestProgress> controller;
-  late StreamSubscription<ProgressEvent> subscription;
-  controller = StreamController<RequestProgress>(onListen: () {
-    subscription = input.listen((ProgressEvent event) {
+  final controller = StreamController<RequestProgress>(sync: true);
+  controller.onListen = () {
+    final subscription = input.listen((ProgressEvent event) {
       controller.add(event.lengthComputable
           ? RequestProgress(event.loaded!, event.total!)
           : RequestProgress());
@@ -33,12 +32,10 @@ StreamTransformer<ProgressEvent, RequestProgress> transformProgressEvents =
         onError: controller.addError,
         onDone: controller.close,
         cancelOnError: cancelOnError);
-  }, onPause: () {
-    subscription.pause();
-  }, onResume: () {
-    subscription.resume();
-  }, onCancel: () {
-    subscription.cancel();
-  });
+    controller
+      ..onPause = subscription.pause
+      ..onResume = subscription.resume
+      ..onCancel = subscription.cancel;
+  };
   return controller.stream.listen(null);
 });